feat classfile and classupload data type
