+++ rasp/jvm/JVMProbe/src/main/java/com/security/smith/client/message/ClassFilter.java
@@ -7,14 +7,13 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 public class ClassFilter {
     private String class_name;
     private String class_path;
-    private String super_class;
     private String class_loader_name;
     private String base_class_name;
     private String base_class_loader_name;
     private int rule_id;
 
     @JsonSerialize(converter = StackTraceConverter.class)
-    private StackTraceElement[] stackTrace;
+    private StackTraceElement[] stack_trace;
 
     public String getClassName() {
         return class_name;
@@ -30,14 +29,6 @@ public class ClassFilter {
 
     public void setClassPath(String class_path) {
         this.class_path = class_path;
-    }
-
-    public String getSuperClass() {
-        return super_class;
-    }
-
-    public void setSuperClass(String super_class) {
-        this.super_class = super_class;
     }
 
     public String getClassLoaderName() {
@@ -73,11 +64,11 @@ public class ClassFilter {
     }
 
     public StackTraceElement[] getStackTrace() {
-        return stackTrace;
+        return stack_trace;
     }
 
     public void setStackTrace(StackTraceElement[] stackTrace) {
-        this.stackTrace = stackTrace;
+        this.stack_trace = stackTrace;
     }
 
 }
